+++ src/Controller/EmplambdaController.php
@@ -14,7 +14,7 @@ use App\Entity\Inscription;
 use App\Form\LoginType;
 
 
-class EmplambdaController extends AbstractController
+class EmpLambdaController extends AbstractController
 {
     #[Route('/emplambda', name: 'app_emplambda')]
     public function index(): Response
@@ -24,70 +24,37 @@ class EmplambdaController extends AbstractController
         ]);
     }
 
-    #[Route('/app_afflambda', name:'app_afflambda')]
+    #[Route('/afflambda', name:'app_afflambda')]
     public function afficheLesformations(Request $request, ManagerRegistry $doctrine, SessionInterface $session)
     {
-        $idEmploye = $session->get('idEmploye');
-        $employe = $doctrine->getManager()->getRepository(Employe::class)->find($idEmploye);
-
-        if(!$session->get('idEmploye')) {
-            $message = "Veuillez vous connecter !";
-            $login = new Employe();
-            $form = $this->createForm(LoginType::class, $login);
-            $form->handleRequest($request);
-            $formations = $doctrine->getManager()->getRepository(Formation::class)->findAll();
-            return $this->render('login/index.html.twig', [
-                'form' => $form->createView(),
-                'message' => $message,
-            ]);
-
-        } else if($employe->getStatut() == 1) {
-            $message = "Veuillez vous connecter à votre compte";
-            $login = new Employe();
-            $form = $this->createForm(LoginType::class, $login);
-            $form->handleRequest($request);
-            $formations = $doctrine->getManager()->getRepository(Formation::class)->findAll();
-            return $this->render('login/index.html.twig', [
-                'form' => $form->createView(),
-                'message' => $message,
-            ]);
-        } else {
-            $employe->getNom();
-            $employe->getPrenom();
-            $formations = $doctrine->getManager()->getRepository(Formation::class)->findAll();
-            foreach($formations as $uneFormation)
-            {
-                $uneFormation->getId();
-                $uneFormation->getDateDebut()->format('Y-m-d');
-                $uneFormation->getNbreHeures();
-                $uneFormation->getDepartement();
-                $uneFormation->getProduit();
-            }
-            if(!$formations)
-            {
-                $message = "Pas de formations";
-            }
-            else {
-                $message = null;
-            }
-            //return $this->render('formations/index.html.twig', ['controller_name' => 'EmplambdaController',]);
-            return $this->render('emplambda/index.html.twig', array('formations' => $formations, 'employe' => $employe, 'message' =>$message));
+        $employe = $this->verifEmploye($session, $doctrine, $request);
+        if (!$employe) {
+            return $this->redirectToRoute('app_login_connexion');
         }
+        
+        $nomEmploye = $employe->getNom();
+        $prenomEmploye = $employe->getPrenom();
+
+        $formations = $doctrine->getManager()->getRepository(Formation::class)->findFormationsNonInscritesPourEmploye($employe);
+        $message = "";
+
+            return $this->render('emplambda/index.html.twig', array('formations' => $formations, 'employe' => $employe, 'message' =>$message));
+        
     }
 
 
-    #[Route('/app_afflambda_inscription/{idForm}', name:'app_afflambda_inscription')]
+    #[Route('/afflambda_inscription/{idForm}', name:'app_afflambda_inscription')]
     public function inscription($idForm, Request $request, ManagerRegistry $doctrine, SessionInterface $session)
     {
+        $employe = $this->verifEmploye($session, $doctrine, $request);
+        if (!$employe) {
+            return $this->redirectToRoute('app_login_connexion');
+        }
+
         $idEmploye = $session->get('idEmploye');
 
         $employe = $doctrine->getManager()->getRepository(Employe::class)->find($idEmploye);
         $formation = $doctrine->getManager()->getRepository(Formation::class)->find($idForm);
-        /*$inscriptions = $doctrine->getManager()->getRepository(Inscription::class)->findBy(['login' => $loginFormulaire]);
-        foreach($inscriptions as $uneInscription)
-        {
-            if($uneInscription->getEmploye() == $idEmploye && $uneInscription->getFormation->getId)
-        }*/
 
         $inscription = new Inscription();
         $inscription->setStatut("En cours");
@@ -98,17 +65,24 @@ class EmplambdaController extends AbstractController
         $entityManager = $doctrine->getManager();
         $entityManager->persist($inscription);
         $entityManager->flush();
+        $message = "Nous avons le plaisir de vous confirmer votre inscription à la formation \"".$formation->getNom(). "\" !";
+        $session->getFlashBag()->add('success', $message);
 
         return $this->redirectToRoute('aff_afflambda_les_inscription');
     }
 
-    #[Route('/aff_afflambda_les_inscription', name:'aff_afflambda_les_inscription')]
-    public function afficheLesInscriptions(ManagerRegistry $doctrine, SessionInterface $session)
+    #[Route('/afflambda_les_inscription', name:'aff_afflambda_les_inscription')]
+    public function afficheLesInscriptions(Request $request, ManagerRegistry $doctrine, SessionInterface $session)
     {
+        $employe = $this->verifEmploye($session, $doctrine, $request);
+        if (!$employe) {
+            return $this->redirectToRoute('app_login_connexion');
+        }
+
         $idEmploye = $session->get('idEmploye');
         $employe = $doctrine->getManager()->getRepository(Employe::class)->find($idEmploye);
         $employe->getPrenom();
-        $inscriptions = $doctrine->getManager()->getRepository(Inscription::class)->findBy(['employe' => $idEmploye]);
+        $inscriptions = $doctrine->getManager()->getRepository(Inscription::class)->findBy(['employe' => $idEmploye, 'statut' => "En cours"]);
         if(!$inscriptions)
         {
             $message = "Pas d'inscriptions";
@@ -116,18 +90,44 @@ class EmplambdaController extends AbstractController
         else {
             $message = null;
         }
+        
         //return $this->render('formations/index.html.twig', ['controller_name' => 'EmplambdaController',]);
         return $this->render('emplambda/inscription.html.twig', array('inscriptions' => $inscriptions , 'employe' => $employe, 'message' =>$message));
     }
 
-    #[Route('/app_inscription_suppr/{idInscription}', name:'app_inscription_suppr')]
-    public function suppFormationAction($idInscription, ManagerRegistry $doctrine)
+    #[Route('/inscription_suppr/{idInscription}', name:'app_inscription_suppr')]
+    public function suppFormationAction($idInscription, Request $request, ManagerRegistry $doctrine, SessionInterface $session)
     {
+        $employe = $this->verifEmploye($session, $doctrine, $request);
+        if (!$employe) {
+            return $this->redirectToRoute('app_login_connexion');
+        }
+
         $formation = $doctrine->getManager()->getRepository(Inscription::class)->find($idInscription);
+        $message = "Votre désinscription de la formation \"". $formation->getFormation()->getNom() . "\" a été enregistrée.";
         $entityManager = $doctrine->getManager();
         $entityManager->remove($formation);
         $entityManager->flush();
+        $session->getFlashBag()->add('success', $message);
+        
         return $this->redirectToRoute('aff_afflambda_les_inscription');
+    }
+
+    private function verifEmploye(SessionInterface $session, ManagerRegistry $doctrine, Request $request): ?Employe
+    {
+        $employe = $session->get('employe');
+
+        if (!$employe || $employe->getStatut() == 1) {
+            $message = "Veuillez vous connecter !";
+            $login = new Employe();
+            $form = $this->createForm(LoginType::class, $login);
+            $form->handleRequest($request);
+            $formations = $doctrine->getManager()->getRepository(Formation::class)->findAll();
+            $this->addFlash('error', $message);
+            return null;
+        }
+
+        return $employe;
     }
